refactor(storybook): dedicated apps/storybook workspace + sound story cache exclusion
Storybook moves out of apps/web into its own workspace, per the Turborepo
Storybook guide's monorepo shape. Stories stay co-located in packages/ui/src;
only the runtime leaves web: .storybook/, the Vitest browser project, the
storybook deps, and the static build. Its dev task is persistent, so root
`pnpm dev` now launches Storybook on :6006 alongside web + api.

This also lands the measured over-invalidation fix as a *sound* pair:

- Tailwind first: ui's globals.css scanned ../**/*.{ts,tsx} including
  stories, so story-only utilities shipped in the apps' CSS — meaning a
  story edit could legitimately change web's build output, and excluding
  stories from web#build's hash would have been unsafe. globals.css now
  excludes *.stories.tsx (`@source not`, via tailwindcss 4.0.8 → 4.3.2)
  and apps/storybook/.storybook/preview.css re-adds the scan where stories
  render. Verified by grepping built CSS: story-only classes present in
  storybook-static, absent from web's .next output; 61/61 storybook
  browser tests green on 4.3.2.
- Then turbo: packages/ui/turbo.json excludes src/**/*.stories.tsx from
  build/transit inputs; apps/storybook/turbo.json folds them back in via
  $TURBO_ROOT$ for build/test:storybook. Proven with --dry=json hash
  diffs: a story edit invalidates only storybook#build/test:storybook
  (web#build/test/typecheck unchanged); a component edit invalidates all.

Rode along:
- Removed two dead @source globs resolving to nonexistent packages/apps
  and packages/components.
- Deleted orphaned packages/config-eslint (only stray node_modules, no
  package.json) and ui's unused @turbo/gen.
- Root scripts use the documented `turbo run <task>` form; new root
  test:storybook delegate; api dev calls `nest start --watch` directly.
- CI storybook job switches to `turbo run test:storybook build
  --filter=storybook` with its own turbo cache key (actionlint + zizmor
  clean).
- oxc-resolver + @oxc-resolver/* excluded from the pnpm release cooldown:
  a storybook transitive publishing ~20 native bindings minutes after each
  frequent parent release, so a fully-mature set rarely exists inside any
  7-day window (documented tradeoff in pnpm-workspace.yaml).
- Root turbo.json comment above `test` rewritten — it still carried the
  disproven "never invalidated" claim corrected in the previous commit.
