+++ CHANGELOG.md
@@ -2,6 +2,7 @@ _Reverse-chronological record of shipped work — features, fixes, and chores. N
 
 # 2026-07-18
 
+- Storybook moves into a dedicated `apps/storybook` workspace (stories stay co-located in `packages/ui/src`; only the runtime — `.storybook/`, the Vitest browser project, the deps, the static build — leaves `apps/web`). Its `dev` task is persistent, so root `pnpm dev` now launches Storybook on :6006 alongside web + api. This also fixes a measured cache-correctness/over-invalidation pair: (1) ui's `globals.css` scanned `../**/*.{ts,tsx}` including stories, so story-only Tailwind utilities shipped in the apps' CSS and any story edit legitimately invalidated `web#build` — stories are now excluded from ui's `@source` scan (`@source not`, via a tailwindcss 4.0.8 → 4.3.2 bump) and scanned instead from the storybook app's own `preview.css`, verified by grepping the built CSS on both sides; (2) with that soundness in place, `packages/ui/turbo.json` excludes `*.stories.tsx` from the `build`/`transit` inputs and `apps/storybook/turbo.json` folds them back in, so a story edit invalidates only `storybook#build`/`test:storybook` while a component edit still invalidates everything (proven by `--dry=json` hash diffs both directions). Rode along: removed two dead `@source` globs pointing at nonexistent `packages/apps`/`packages/components`; deleted the orphaned `packages/config-eslint` directory (no `package.json`, only stray `node_modules`); dropped ui's unused `@turbo/gen`; root scripts now use the documented `turbo run <task>` form; api `dev` calls `nest start --watch` directly; CI's storybook job switches to `turbo run test:storybook build --filter=storybook` with its own turbo cache; `oxc-resolver` + `@oxc-resolver/*` (storybook transitive with ~20 co-published native bindings, several releases/month) are excluded from the pnpm release cooldown because a fully-mature binding set rarely exists inside any 7-day window. All 61 storybook browser tests pass on Tailwind 4.3.2.
 - Turborepo: adopt the documented transit-node pattern and vendor the Turborepo Claude skill. `typecheck` was serialized (`dependsOn: ["^typecheck"]`) and `test` carried an inaccurate `dependsOn: ["^build"]` (tests compile from source; no package emits build output they consume). Both now depend on a scriptless `transit` node (`transit: { dependsOn: ["^transit"] }`), so they run in parallel across packages while still cache-busting when an imported workspace package changes — verified a bare `{}` (no edge) drops that invalidation, so the `^`-traversal edge is load-bearing (the original `^`-based config was already correct on this axis; this is a parallelism + intent change, not a correctness fix). Also excluded `!.next/dev/**` from the web build outputs, vendored `.agents/skills/turborepo`, and excluded `.agents/` (third-party skill content) from prettier.
 - Package-manager supply-chain hardening. Added a `minimumReleaseAge: 10080` (7-day) release cooldown in `pnpm-workspace.yaml` — pnpm refuses to _resolve_ a version until it has aged a week, so a compromised release has a wide detection window before it can enter the lockfile (`@typescript/native-preview`, which tracks fresh dated dev builds, is excluded; `--frozen-lockfile` CI installs skip re-resolution and are unaffected, so only manual `pnpm add`/`update` are gated). Added a `preinstall: npx only-allow@1.2.1 pnpm` guard (version-pinned to an immutable, long-stable release so the guard itself can't be tarball-swapped at install time) that rejects `npm`/`yarn` installs (which would ignore the `allowBuilds` script-blocklist and the catalog), and `engine-strict=true` (`.npmrc`) so a Node version below the `>=22.12` floor fails the install instead of only warning. Complements the existing pinned `packageManager`, `allowBuilds` lifecycle-script allowlist, and frozen-lockfile CI. Tooling only.
 - Repo-wide prettier + lint gating (closing **#175**). The root `format:check` only covered `apps/api/{src,test}/**/*.ts`, so apps/web, packages/ui, and all Markdown/JSON/YAML/CSS drifted un-gated; the `format` glob had no ignore file and descended into the ~5k files under `.claude/worktrees/`. `format`/`format:check` are now a single ignore-pruned, cached `prettier --write|--check .`, backed by a new `.prettierignore` (worktrees, build output, local gitignored files, `pnpm-lock.yaml`). Generated artifacts auto-format at their source so they can be gated without coupling CI to a generator: the api `build` prettier-writes `openapi.json` and `db:generate` prettier-writes the drizzle `meta/` JSON (`drizzle-kit check`/`generate` verified clean over the reformat). The lefthook pre-commit `format` job goes repo-wide over staged prettier-parseable files (keeping the #173 ACMR-deletion guard), the ui lint glob widens to `{ts,tsx,js,mjs,cjs}` so `*.mjs` configs are linted pre-commit, and turbo's needless `^lint` edge is dropped so lint jobs run fully parallel. Plus a one-time mechanical reflow of the newly-covered surface. Tooling only; no runtime behavior change.
