+++ apps/storybook/.oxlintrc.json
@@ -1,0 +1,13 @@
+{
+  "$schema": "./node_modules/oxlint/configuration_schema.json",
+  "plugins": ["typescript", "unicorn", "oxc", "react"],
+  "env": {
+    "browser": true,
+    "node": true
+  },
+  "ignorePatterns": ["storybook-static/**"],
+  "rules": {
+    "react/rules-of-hooks": "error",
+    "react/exhaustive-deps": "warn"
+  }
+}
