+++ apps/storybook/globals.d.ts
@@ -1,0 +1,3 @@
+// Side-effect CSS imports (preview.css). apps/web gets this from Next's own
+// global types via next-env.d.ts; this package has no Next entrypoint.
+declare module "*.css";
