+++ apps/storybook/postcss.config.mjs
@@ -1,0 +1,1 @@
+export { default } from "@workspace/ui/postcss.config";
