+++ package.json
@@ -5,11 +5,12 @@
   "scripts": {
     "preinstall": "npx only-allow@1.2.1 pnpm",
     "prepare": "lefthook install",
-    "build": "turbo build",
-    "dev": "turbo dev",
-    "lint": "turbo lint",
-    "test": "turbo test",
-    "typecheck": "turbo typecheck",
+    "build": "turbo run build",
+    "dev": "turbo run dev",
+    "lint": "turbo run lint",
+    "test": "turbo run test",
+    "test:storybook": "turbo run test:storybook",
+    "typecheck": "turbo run typecheck",
     "format": "prettier --write --cache .",
     "format:check": "prettier --check --cache .",
     "test:e2e": "playwright test",
