+++ packages/ui/package.json
@@ -31,8 +31,7 @@
   },
   "devDependencies": {
     "@storybook/nextjs-vite": "catalog:",
-    "@tailwindcss/postcss": "^4.0.8",
-    "@turbo/gen": "^2.4.2",
+    "@tailwindcss/postcss": "^4.1.0",
     "@types/node": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
@@ -40,7 +39,7 @@
     "@workspace/config-typescript": "workspace:*",
     "oxlint": "catalog:",
     "storybook": "catalog:",
-    "tailwindcss": "^4.0.8",
+    "tailwindcss": "^4.1.0",
     "typescript": "catalog:"
   },
   "exports": {
